added frontend starter files

+++ backend/controller/product.controller.js
@@ -29,7 +29,10 @@ export const createProduct =  async (req, res)=>{
 }
 
 export const deleteProduct =  async (req, res)=>{
-    const productId = req.params;
+    const {id} = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success:false, message:"Invalid Product ID"})
+    }
     try{
         await Product.findByIdAndDelete(id);
         res.status(200).json({success:true, message:"Product deleted"});
